refactor(cli): extract status command into commands/session_commands.py
Move the inline `dopemux status` diagnostic HUD and its
`_get_attention_emoji` helper out of cli.py into a dedicated
session_commands module. ConfigManager is imported at module scope so
mock.patch resolves correctly in tests.

Updates the status tests in test_cli.py to patch the new import site
(dopemux.commands.session_commands.*) and updates the helper's unit
test to import from the new location.
